configured Elmah logging to the DB for functional tests project refactoring of form events (handlers moved to form provider) fixed exposed constants on form events
