+++ shesha-functional-tests/backend/test/Boxfusion.SheshaFunctionalTests.Common.Domain.Tests/Boxfusion.SheshaFunctionalTests.Common.Domain.Tests.csproj
@@ -29,7 +29,7 @@
 		<PackageReference Include="NUnit3TestAdapter" Version="4.5.0" />
 	</ItemGroup>
 	<ItemGroup>
-		<PackageReference Include="Microsoft.Identity.Client" Version="4.58.1" />
+		<PackageReference Include="Microsoft.Identity.Client" Version="4.60.3" />
 		<PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.8.0" />
 		<PackageReference Include="Moq" Version="4.20.70" />
 		<PackageReference Include="NSubstitute" Version="5.1.0" />
