+++ shesha-reactjs/src/components/configurableForm/configurableFormRenderer.tsx
@@ -1,55 +1,38 @@
 import _ from 'lodash';
-import axios, { AxiosResponse } from 'axios';
 import classNames from 'classnames';
-import cleanDeep from 'clean-deep';
 import ComponentsContainer from '../formDesigner/containers/componentsContainer';
-import moment from 'moment';
-import qs from 'qs';
 import React, {
   FC,
   PropsWithChildren,
   useEffect,
   useMemo,
   useState
-  } from 'react';
+} from 'react';
 import {
-  addFormFieldsList,
   hasFiles,
   jsonToFormData,
-  removeGhostKeys
-  } from '@/utils/form';
-import { axiosHttp } from '@/utils/fetchers';
+} from '@/utils/form';
 import { ComponentsContainerForm } from '../formDesigner/containers/componentsContainerForm';
 import { ComponentsContainerProvider } from '@/providers/form/nesting/containerContext';
-import { Form, message, Spin } from 'antd';
-import { filterDataByOutputComponents, FormConfigurationDto, useFormData } from '@/providers/form/api';
+import { Form, Spin } from 'antd';
+import { useFormData } from '@/providers/form/api';
 import { getQueryParams } from '@/utils/url';
-import { IAbpWrappedGetEntityResponse } from '@/interfaces/gql';
-import { IAnyObject, IDictionary, ValidateErrorEntity } from '@/interfaces';
-import { IConfigurableFormRendererProps, IDataSourceComponent } from './models';
-import { nanoid } from '@/utils/uuid';
+import { IAnyObject, ValidateErrorEntity } from '@/interfaces';
+import { IConfigurableFormRendererProps } from './models';
 import { ROOT_COMPONENT_KEY } from '@/providers/form/models';
 import { StandardEntityActions } from '@/interfaces/metadata';
-import { useDataContextManager } from '@/providers/dataContextManager/index';
-import { useDelayedUpdate } from '@/providers/delayedUpdateProvider';
 import { useForm } from '@/providers/form';
 import { useFormDesigner } from '@/providers/formDesigner';
-import { useGlobalState, useSheshaApplication } from '@/providers';
+import { useGlobalState } from '@/providers';
 import { useModelApiEndpoint } from './useActionEndpoint';
 import { useMutate } from '@/hooks/useMutate';
 import { useStyles } from './styles/styles';
 import {
   evaluateComplexString,
-  evaluateKeyValuesToObjectMatchedData,
-  evaluateString,
   evaluateValue,
-  executeScript,
-  getComponentNames,
   getObjectWithOnlyIncludedKeys,
   IMatchData,
 } from '@/providers/form/utils';
-import { useFormDesignerComponents } from '@/providers/form/hooks';
-import { SheshaCommonContexts } from '@/providers/dataContextManager/models';
 
 export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRendererProps>> = ({
   children,
@@ -65,9 +48,6 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
 
   const formInstance = useForm();
   const { styles } = useStyles();
-  //const contextManager = useDataContextManager(false);
-  //if (contextManager)
-  //  contextManager.updateFormInstance(formInstance);
 
   const {
     updateStateFormData,
@@ -84,14 +64,9 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
   const designerMode = formMode === 'designer';
 
   const { isDragging = false } = useFormDesigner(false) ?? {};
-  const toolboxComponents = useFormDesignerComponents();
-  
 
-  const { excludeFormFieldsInPayload, onDataLoaded, onUpdate, onInitialized, formKeysToPersist, uniqueFormId } =
-    formSettings;
-  const { globalState, setState: setGlobalState } = useGlobalState();
-  const dcm = useDataContextManager(false);
-  const pageContext = dcm.getPageContext();
+  const { onDataLoaded, onUpdate, onInitialized, formKeysToPersist, uniqueFormId } = formSettings;
+  const { globalState } = useGlobalState();
 
   const urlEvaluationData: IMatchData[] = [
     { match: 'initialValues', data: initialValues },
@@ -107,7 +82,6 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
     mappings: urlEvaluationData,
   });
 
-  const { backendUrl, httpHeaders } = useSheshaApplication();
   const [lastTruthyPersistedValue, setLastTruthyPersistedValue] = useState<IAnyObject>(null);
 
   const dataFetcher = useFormData({
@@ -118,79 +92,13 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
     // data: formData, NOTE: form data must not be used here!
   });
 
-  const { getPayload: getDelayedUpdate } = useDelayedUpdate(false) ?? {};
-
   const queryParamsFromAddressBar = useMemo(() => getQueryParams(), []);
-
-  const sheshaUtils = {
-    prepareTemplate: (templateId: string, replacements: object): Promise<string> => {
-      if (!templateId) return Promise.resolve(null);
-
-      const payload = {
-        id: templateId,
-        properties: 'markup',
-      };
-      const url = `${backendUrl}/api/services/Shesha/FormConfiguration/Query?${qs.stringify(payload)}`;
-      return axios
-        .get<any, AxiosResponse<IAbpWrappedGetEntityResponse<FormConfigurationDto>>>(url, { headers: httpHeaders })
-        .then((response) => {
-          const markup = response.data.result.markup;
-
-          const preparedMarkup = evaluateString(markup, {
-            NEW_KEY: nanoid(),
-            GEN_KEY: nanoid(),
-            ...(replacements ?? {}),
-          });
-
-          return preparedMarkup;
-        });
-    },
-  };
-
-  const executeExpression = (
-    expression: string,
-    includeInitialValues = true,
-    includeMoment = true,
-    includeAxios = true,
-    includeMessage = true,
-    exposedData = null
-  ) => {
-    if (!expression) {
-      return null;
-    }
-
-    const application = dcm?.getDataContext(SheshaCommonContexts.ApplicationContext);
-
-    const callArguments: IDictionary<any> = {
-      data: exposedData || formData, 
-      parentFormValues: parentFormValues, 
-      globalState: globalState, 
-      shesha: sheshaUtils, 
-      form: form, 
-      setFormData: setFormData, 
-      setGlobalState: setGlobalState, 
-      contexts: { ...dcm?.getDataContextsData(), lastUpdate: dcm?.lastUpdate }, 
-      application: application?.getData(),
-      pageContext: pageContext?.getFull(),
-    };
-    if (includeInitialValues)
-      callArguments.initialValues = initialValues;
-    if (includeMoment)
-      callArguments.moment = moment;
-    if (includeAxios)
-      callArguments.http = axiosHttp(backendUrl);
-    if (includeMessage)
-      callArguments.message = message;
-
-    return executeScript(expression, callArguments);
-  };
-
 
   // Execute onInitialized if provided
   useEffect(() => {
     if (onInitialized) {
       try {
-        executeExpression(onInitialized);
+        formInstance.executeExpression(onInitialized);
       } catch (error) {
         console.warn('onInitialized error', error);
       }
@@ -254,13 +162,13 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
 
   useEffect(() => {
     if (fetchedFormEntity && onDataLoaded) {
-      executeExpression(onDataLoaded, true, true, true, true, fetchedFormEntity); // On Initialize
+      formInstance.executeExpression(onDataLoaded, fetchedFormEntity); // On Initialize
     }
   }, [onDataLoaded, fetchedFormEntity]);
 
   useEffect(() => {
     if (onUpdate) {
-      executeExpression(onUpdate); // On Update
+      formInstance.executeExpression(onUpdate); // On Update
     }
   }, [formData, onUpdate]);
 
@@ -304,91 +212,14 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
 
   const { mutate: doSubmit, loading: submitting } = useMutate();
 
-  const getDynamicPreparedValues = (): Promise<object> => {
-    const { preparedValues } = formSettings;
-
-    return Promise.resolve(preparedValues ? executeExpression(preparedValues) : {});
-  };
-
-  const getInitialValuesFromFormSettings = () => {
-    const initialValuesFromFormSettings = formSettings?.initialValues;
-
-    const values = evaluateKeyValuesToObjectMatchedData(initialValuesFromFormSettings, [
-      { match: 'data', data: formData },
-      { match: 'parentFormValues', data: parentFormValues },
-      { match: 'globalState', data: globalState },
-    ]);
-
-    return cleanDeep(values, {
-      // cleanKeys: [], // Don't Remove specific keys, ie: ['foo', 'bar', ' ']
-      // cleanValues: [], // Don't Remove specific values, ie: ['foo', 'bar', ' ']
-      // emptyArrays: false, // Don't Remove empty arrays, ie: []
-      // emptyObjects: false, // Don't Remove empty objects, ie: {}
-      // emptyStrings: false, // Don't Remove empty strings, ie: ''
-      // NaNValues: true, // Remove NaN values, ie: NaN
-      // nullValues: true, // Remove null values, ie: null
-      undefinedValues: true, // Remove undefined values, ie: undefined
-    });
-  };
-
   const options = { setValidationErrors };
 
   const convertToFormDataIfRequired = (data: any) => {
     return data && hasFiles(data) ? jsonToFormData(data) : data;
   };
 
-  const prepareDataForSubmit = (): Promise<object> => {
-    const initialValuesFromFormSettings = getInitialValuesFromFormSettings();
-
-    return getDynamicPreparedValues()
-      .then((dynamicValues) => {
-        const initialValues = getInitialValuesFromFormSettings();
-        const nonFormValues = { ...dynamicValues, ...initialValues };
-
-        let postData = excludeFormFieldsInPayload
-          ? removeGhostKeys({ ...formData, ...nonFormValues })
-          : removeGhostKeys(addFormFieldsList(formData, nonFormValues, form));
-
-        postData = filterDataByOutputComponents(postData, allComponents, toolboxComponents);
-
-        const delayedUpdate = typeof getDelayedUpdate === 'function' ? getDelayedUpdate() : null;
-        if (Boolean(delayedUpdate)) postData._delayedUpdate = delayedUpdate;
-
-        const subFormNamesToIgnore = getComponentNames(
-          allComponents,
-          (component: IDataSourceComponent) =>
-            (component?.type === 'list' || component?.type === 'subForm') && component.dataSource === 'api'
-        );
-
-        if (subFormNamesToIgnore?.length) {
-          subFormNamesToIgnore.forEach((key) => {
-            if (Object.hasOwn(postData, key)) {
-              delete postData[key];
-            }
-          });
-          const isEqualOrStartsWith = (input: string) =>
-            subFormNamesToIgnore?.some((x) => x === input || input.startsWith(`${x}.`));
-
-          postData._formFields = postData._formFields?.filter((x) => !isEqualOrStartsWith(x));
-        }
-
-        if (excludeFormFieldsInPayload) {
-          delete postData._formFields;
-        } else {
-          if (initialValuesFromFormSettings) {
-            postData._formFields = Array.from(
-              new Set<string>([...(postData._formFields || []), ...Object.keys(initialValuesFromFormSettings)])
-            );
-          }
-        }
-
-        return postData;
-      })
-      .catch((error) => console.error(error));
-  };
-
   const onFinishInternal = () => {
-    prepareDataForSubmit().then((postData) => {
+    formInstance.prepareDataForSubmit().then((postData) => {
       if (props.onFinish) {
         props.onFinish(postData, options);
       } else {
@@ -454,7 +285,7 @@ export const ConfigurableFormRenderer: FC<PropsWithChildren<IConfigurableFormRen
         {...mergedProps}
       >
         <ComponentsContainerProvider ContainerComponent={ComponentsContainerForm}>
-          <ComponentsContainer containerId={ROOT_COMPONENT_KEY}   />
+          <ComponentsContainer containerId={ROOT_COMPONENT_KEY} />
         </ComponentsContainerProvider>
         {children}
       </Form>
